refactor: add VideoDurationExtractorImpl for android and ios
