+++ trends_presentation/src/androidMain/kotlin/net/thechance/mena/trends/presentation/shared/util/video_util/VideoDurationExtractor.android.kt
@@ -1,0 +1,40 @@
+package net.thechance.mena.trends.presentation.shared.util.video_util
+
+import android.media.MediaMetadataRetriever
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import java.io.File
+import java.io.FileOutputStream
+
+actual fun getVideoDurationExtractor(): VideoDurationExtractor {
+    return VideoDurationExtractorImpl()
+}
+
+class VideoDurationExtractorImpl : VideoDurationExtractor {
+    override suspend fun getDuration(videoBytes: ByteArray): Long? {
+        return withContext(Dispatchers.IO) {
+            val retriever = MediaMetadataRetriever()
+            var tempFile: File? = null
+
+            try {
+                tempFile = File.createTempFile("temp_video", ".mp4")
+                FileOutputStream(tempFile).use { it.write(videoBytes) }
+
+                retriever.setDataSource(tempFile.absolutePath)
+                val durationStr = retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_DURATION)
+
+                durationStr?.toLongOrNull()
+            } catch (e: Exception) {
+                e.printStackTrace()
+                null
+            } finally {
+                try {
+                    retriever.release()
+                    tempFile?.delete()
+                } catch (e: Exception) {
+                    e.printStackTrace()
+                }
+            }
+        }
+    }
+}
