+++ trends_presentation/src/commonMain/kotlin/net/thechance/mena/trends/presentation/shared/util/video_util/VideoDurationExtractor.kt
@@ -1,5 +1,7 @@
 package net.thechance.mena.trends.presentation.shared.util.video_util
 
+expect fun getVideoDurationExtractor(): VideoDurationExtractor
+
 interface VideoDurationExtractor {
     suspend fun getDuration(videoBytes: ByteArray): Long?
 }
