+++ trends_presentation/src/commonMain/kotlin/net/thechance/mena/trends/presentation/shared/util/video_util/videoUtils.kt
@@ -2,8 +2,6 @@ package net.thechance.mena.trends.presentation.shared.util.video_util
 
 import kotlin.math.roundToInt
 
-expect suspend fun getVideoDuration(videoBytes: ByteArray): Long?
-
 fun formatBytes(bytes: Long): String {
     val kb = 1024.0
     val mb = kb * 1024
